% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-detailpage.R
\name{calc_price_summary}
\alias{calc_price_summary}
\title{Calculate a summary of prices in product detail page}
\usage{
calc_price_summary(detailpagehtml)
}
\arguments{
\item{detailpagehtml}{html structure from a single geizhals page
listing details of a specific item.}
}
\value{
A tibble (data.frame) with two columns (key and value),
  containing the price summary results (key being a descriptive
  key like \code{price_min}, value being the respective summary
  measure of the prices). The value column is of type
  \code{character}, in order to be row-binded to the categories
  and their values (which are also of type \code{character}).
}
\description{
Returns a summary of all price values from the price list
in the detailed product description page. Currently,
this summary contains the 3 lowest prices (or \code{NA} if
there aren't enough prices on that page), and the median
of all prices.
}
\examples{
\dontrun{
## get data from multiple geizhals category pages:
url_geizhals <- "https://geizhals.at/?cat=acam35"
listpagehtml_list <- fetch_all_listpages(url_geizhals, max_pages = 2)
dat_listpage <- parse_all_listpages(listpagehtml_list)
## get url of a single detail page and read html:
url_detailpage <- dat_listpage[["detailpage_url"]][1]
detailpagehtml <- xml2::read_html(url_detailpage)
## get prices summary:
calc_price_summary(detailpagehtml)
}

}
