% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-geizhals.R
\name{extract_domain}
\alias{extract_domain}
\title{Extract the domain from the full URL}
\usage{
extract_domain(url_str)
}
\arguments{
\item{url_str}{URL of some (geizhals) page.
Character vector of length 1.}
}
\value{
The domain up to (but not including) the last "/"
  character, as a character vector of length 1.
}
\description{
Extracts the domain (usually \code{https://geizhals.at} or
\code{https://geizhals.at}, but may also be the path to a
local file for testing or debugging.)
}
\examples{
extract_domain("https://geizhals.eu/?cat=monlcd19wide")
extract_domain(paste0("https://geizhals.at/?cat=hwaeschtr&",
  "xf=1027_W\%E4rmepumpentrockner\%7E1296_10\%7E1747_8\%7E7641_40\%7E7653_9"))
extract_domain(paste0("https://geizhals.at/lg-electronics-x-",
  "screen-k500n-weiss-a1419167.html?hloc=at"))
extract_domain(system.file("extdata", "gh-lst-trockn-01.html",
  package = "rgeizhals"))

}
