% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-detailpage.R
\name{get_all_detailpages}
\alias{get_all_detailpages}
\title{Get data from multiple product detail pages}
\usage{
get_all_detailpages(detailpagehtml_list)
}
\arguments{
\item{detailpagehtml_list}{A list of html structure
from multiple geizhals page listing details of a
specific item.}
}
\value{
A tibble (data.frame) with as many columns
  as there are distinct categories in all feature
  pages, and as many rows as there are products.
}
\description{
Returns all categories and their values in a list of
detailed product description pages, as well as
a summary of all price values from the price list
in each of the  detailed product description pages.
In contrast to the \code{get_single_detailpage}
function, the categories describing a product are
the columns, and each product is represented as a
row in the resulting tibble (data.frame).
The tibble has as many columns as there are categories,
if a product doesn't feature all categories in its
description, this column will be \code{NA}.
}
\examples{

## get data from multiple geizhals category pages:
url_geizhals <- "https://geizhals.at/?cat=acam35"
listpagehtml_list <- read_all_listpages(url_geizhals, max_pages = 2)
dat_listpage <- get_all_listpages(listpagehtml_list)
## pick only the three first detailpage urls:
wch_detailpage_urls <- dat_listpage[["detailpage_url"]][1:3]
detailpagehtml_list <- read_all_detailpage_html(wch_detailpage_urls)
## get data from all detailpages:
dat_detailpages <- get_all_detailpages(detailpagehtml_list)
head(dat_detailpages)

}
