% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-geizhals.R
\name{get_feature_summary}
\alias{get_feature_summary}
\title{Get summary table for some feature}
\usage{
get_feature_summary(dat_gh, col, sep = ",")
}
\arguments{
\item{dat_gh}{A tibble (data.frame) containing the data.}

\item{col}{The column name that should be summarized
(character vector of length 1).}

\item{sep}{The separator to use as split, if the column contains
a list of features. Character vector of length 1. Default is ",".}
}
\value{
A frequency table of features.
}
\description{
Produces a frequency table of all features in a specific column.
Especially useful if one column contains a list of features.
}
\examples{
\dontrun{
url_geizhals <- "https://geizhals.at/?cat=acam35"
dat_gh <- get_geizhals_data(url_geizhals, max_pages = 1)
get_feature_summary(dat_gh, col = "Typ")
}

}
