% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-listpage.R
\name{get_next_listpage_url}
\alias{get_next_listpage_url}
\title{Get url of next listpage}
\usage{
get_next_listpage_url(listpagehtml)
}
\arguments{
\item{listpagehtml}{html structure from a single geizhals page of search
results as gathered via \code{xml2::read_html()} or via a single entry
of the list of search pages resulting from \code{read_all_listpages}.}
}
\value{
Character vector of length 1 with the full url of the next page
  with product listings, or \code{NA} if no more page is available.
}
\description{
Returns the url of the next page of listings, or \code{NA} if no more
page is availab.e
}
\examples{

}
