% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-geizhals.R
\name{join_details_to_listpage}
\alias{join_details_to_listpage}
\title{Join details to listpage data}
\usage{
join_details_to_listpage(dat_listpage, dat_detailpage)
}
\arguments{
\item{dat_listpage}{A tibble (data.frame) containing
all the data from the scraped geizhals category
listing pages, with the join column
\code{detailpage_url}.}

\item{dat_detailpage}{A tibble (data.frame) containing
the data from corresponding detail pages, with the
column specifying the \code{url}.}
}
\value{
A tibble (data.frame) containing both the data
  from the category pages as well as the corresponding
  detail page data (or \code{NA} if no match is present).
}
\description{
Joins the data from the product detail pages to
the data from the category pages.
}
\examples{
\dontrun{
url_geizhals <- "https://geizhals.at/?cat=acam35"
## fetch html of all listing pages:
listpagehtml_list <- fetch_all_listpages(url_geizhals, max_pages = 2)
## and parse information of these listing pages:
dat_listpage <- parse_all_listpages(listpagehtml_list)
## get all (or some) detailpages:
detailpagehtml_list <- fetch_all_detailpage_html(dat_listpage$detailpage_url,
                                                 max_items = 5)
dat_detailpage <- parse_all_detailpages(detailpagehtml_list)
dat_geizhals <- join_details_to_listpage(dat_listpage,
                                         dat_detailpage)
head(dat_geizhals)
}
}
