% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-listpage.R
\name{parse_all_listpages}
\alias{parse_all_listpages}
\title{Parse data from multiple geizhals category pages}
\usage{
parse_all_listpages(listpagehtml_list, domain)
}
\arguments{
\item{listpagehtml_list}{List of html structures from a multiple
geizhals pages, each listing items in a selected category,
as gathered via \code{fetch_all_listpages}.}

\item{domain}{Character vector of length one specifying the domain.
Defaults to \code{"https://geizhals.at"}.}
}
\value{
A tibble (data.frame) containing all information scraped
  from the geizhals pages.
}
\description{
Takes a list of html as input, and returns the information in
tabluar form.
}
\examples{
\dontrun{
url_geizhals <- "https://geizhals.at/?cat=acam35"
listpagehtml_list <- fetch_all_listpages(url_geizhals, max_pages = 2)
parse_all_listpages(listpagehtml_list)
}

}
