% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-detailpage.R
\name{parse_keyval_tbl}
\alias{parse_keyval_tbl}
\title{Parse categories and their values in product detail page}
\usage{
parse_keyval_tbl(detailpagehtml)
}
\arguments{
\item{detailpagehtml}{html structure from a single geizhals page
listing details of a specific item.}
}
\value{
A tibble (data.frame) with two columns (key and value),
  containing the categories and their values.
}
\description{
Returns all categories (titles) and their values
in the detailed product description page. The categories
are returned in the order that they appear in on the page,
and the categories might not be identical on the detailed
product description pages of different products within
the same category.
}
\examples{
\dontrun{
## get data from multiple geizhals category pages:
url_geizhals <- "https://geizhals.at/?cat=acam35"
listpagehtml_list <- fetch_all_listpages(url_geizhals, max_pages = 2)
dat_listpage <- parse_all_listpages(listpagehtml_list)
## get url of a single detail page and read html:
url_detailpage <- dat_listpage[["detailpage_url"]][1]
detailpagehtml <- xml2::read_html(url_detailpage)
## get categories and their values:
parse_keyval_tbl(detailpagehtml)
}

}
