% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-listpage.R
\name{parse_listprice}
\alias{parse_listprice}
\title{Parse price from listing page}
\usage{
parse_listprice(listpagehtml)
}
\arguments{
\item{listpagehtml}{html structure from a single geizhals page listing
items in a selected category, as gathered via \code{xml2::read_html()}
or via a single entry of the list of search pages resulting from
\code{fetch_all_listpages}.}
}
\value{
A numeric vector containing the price in the same
  order of the products as appearing in the listing page
  returned from \code{parse_product_names}.
}
\description{
Returns the listed price for a product in a single geizhals
html page of a category listing page. The order
might not correspond to the order listed on the webpage, but it is
the same order in all related functions.
}
\examples{
\dontrun{
## get html of a geizhals category page via read_html():
url_geizhals <- "https://geizhals.at/?cat=acam35"
listpagehtml <- xml2::read_html(url_geizhals)
parse_listprice(listpagehtml)

## get html of multiple geizhals search pages:
listpagehtml_list <- fetch_all_listpages(url_geizhals, max_pages = 2)
parse_listprice(listpagehtml_list[[1]])
}

}
