% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-listpage.R
\name{parse_offers_n}
\alias{parse_offers_n}
\title{Parse numbers of offers from geizhals listing page}
\usage{
parse_offers_n(listpagehtml)
}
\arguments{
\item{listpagehtml}{html structure from a single geizhals page listing
items in a selected category, as gathered via \code{xml2::read_html()}
or via a single entry of the list of listing pages resulting from
\code{fetch_all_listpages}.}
}
\value{
A numeric vector containing the number of offers in the same
  order of the products as appearing in the category listing page
  returned from \code{parse_product_names}.
}
\description{
Returns the numbers of offers for a product in a single geizhals
html page of category listing results. The order
might not correspond to the order listed on the webpage, but it is
the same order in all related functions.
}
\examples{
\dontrun{
## get html of a geizhals category listing page via read_html:
url_geizhals <- "https://geizhals.at/?cat=acam35"
listpagehtml <- xml2::read_html(url_geizhals)
parse_offers_n(listpagehtml)

## get html of multiple geizhals category listing pages:
listpagehtml_list <- fetch_all_listpages(url_geizhals)
parse_offers_n(listpagehtml_list[[1]])
}

}
