% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-detailpage.R
\name{parse_prices}
\alias{parse_prices}
\title{Get price list in product detail page}
\usage{
parse_prices(detailpagehtml)
}
\arguments{
\item{detailpagehtml}{html structure from a single geizhals page
listing details of a specific item.}
}
\value{
A numeric vector containing the prices.
}
\description{
Returns all price values from the price list
in the detailed product description page.
}
\examples{
\dontrun{
## get data from multiple geizhals category pages:
url_geizhals <- "https://geizhals.at/?cat=acam35"
listpagehtml_list <- fetch_all_listpages(url_geizhals, max_pages = 2)
dat_listpage <- parse_all_listpages(listpagehtml_list)
## get url of a single detail page and read html:
url_detailpage <- dat_listpage[["detailpage_url"]][1]
detailpagehtml <- xml2::read_html(url_detailpage)
## get prices:
parse_prices(detailpagehtml)
}

}
