% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeizhals-package.R
\docType{package}
\name{rgeizhals}
\alias{rgeizhals}
\alias{rgeizhals-package}
\title{\code{rgeizhals} package}
\description{
Scrapes www.geizhals.at pages
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ingonader/rgeizhals}
  \item Report bugs at \url{https://github.com/ingonader/rgeizhals/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
