% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeizhals-package.R
\docType{package}
\name{rgeizhals}
\alias{rgeizhals}
\alias{rgeizhals-package}
\title{\code{rgeizhals} package}
\description{
Scrapes www.geizhals.at pages
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
